requirements.txt
streamlit
cohere
requests
sentence-transformers
faiss-cpu
